+++ lzx/encoder/encapi.h
@@ -16,16 +16,16 @@
 bool LZX_EncodeInit(
 
 	t_encoder_context *	context,
-	long				compression_window_size,
-	long				second_partition_size,
+	int				compression_window_size,
+	int				second_partition_size,
 	PFNALLOC			pfnma,
 	PFNFREE				pfnmf,
 
 	int FAR (DIAMONDAPI *pfnlzx_output_callback)(
 			void *			pfol,
 			unsigned char *	compressed_data,
-			long			compressed_size,
-			long			uncompressed_size
+			int			compressed_size,
+			int			uncompressed_size
     ),
 
     void FAR *          fci_data
@@ -35,19 +35,19 @@ void LZX_EncodeFree(t_encoder_context *context);
 
 void LZX_EncodeNewGroup(t_encoder_context *context);
 
-long LZX_Encode(
+int LZX_Encode(
 	t_encoder_context *	context,
 	byte *				input_data,
-	long				input_size,
-	long *				bytes_compressed,
-	long				file_size_for_translation
+	int				input_size,
+	int *				bytes_compressed,
+	int				file_size_for_translation
 );
 
 bool LZX_EncodeFlush(t_encoder_context *context);
 
 unsigned char *LZX_GetInputData(
     t_encoder_context *context,
-    unsigned long *input_position,
-    unsigned long *bytes_available
+    unsigned int *input_position,
+    unsigned int *bytes_available
 );
